+++ lib/theme_Personal_Info/statitics.dart
@@ -1,0 +1,130 @@
+import 'package:flutter/material.dart';
+
+class StatisticsWidget extends StatelessWidget {
+  const StatisticsWidget({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+      height: 60,
+      color: Colors.white,
+      child: Row(
+        key: const Key("statistics"),
+        mainAxisAlignment: MainAxisAlignment.spaceEvenly,
+        children: [
+          Expanded(
+            child: Column(
+              children: [
+                Expanded(
+                  child: Row(
+                    // mainAxisAlignment: MainAxisAlignment.center,
+                    children: [
+                      const Icon(Icons.attach_money,
+                          size: 40, color: Colors.black),
+                      Column(
+                        mainAxisAlignment: MainAxisAlignment.center,
+                        children: const [
+                          Text(
+                            'Tổng đã chi',
+                            style: TextStyle(
+                                color: Colors.black,
+                                fontSize: 14,
+                                fontWeight: FontWeight.bold),
+                          ),
+                          SizedBox(height: 4.0),
+                          Text(
+                            '100.000.000đ',
+                            style: TextStyle(color: Colors.black),
+                          ),
+                        ],
+                      )
+                    ],
+                  ),
+                )
+              ],
+            ),
+          ),
+          Container(
+            // color: Colors.white,
+            padding: const EdgeInsets.symmetric(vertical: 7),
+            child: const VerticalDivider(
+              thickness: 0.5,
+              color: Colors.black,
+            ),
+          ),
+          Expanded(
+            child: Column(
+              children: [
+                Expanded(
+                  child: Row(
+                    children: [
+                      const Icon(Icons.attach_money,
+                          size: 40, color: Colors.black),
+                      Column(
+                        mainAxisAlignment: MainAxisAlignment.center,
+                        children: const [
+                          Text(
+                            'Tổng đã chi',
+                            style: TextStyle(
+                                color: Colors.black,
+                                fontSize: 14,
+                                fontWeight: FontWeight.bold),
+                          ),
+                          SizedBox(height: 4.0),
+                          Text(
+                            '100.000đ',
+                            style: TextStyle(color: Colors.black),
+                          ),
+                        ],
+                      )
+                    ],
+                  ),
+                )
+              ],
+            ),
+          ),
+          Container(
+            // color: Colors.white,
+            padding: const EdgeInsets.symmetric(vertical: 7),
+            child: const VerticalDivider(
+              thickness: 0.5,
+              color: Colors.black,
+            ),
+          ),
+          Expanded(
+            child: Column(
+              children: [
+                Expanded(
+                  child: Row(
+                    // mainAxisAlignment: MainAxisAlignment.center,
+                    children: [
+                      const Icon(Icons.attach_money,
+                          size: 40, color: Colors.black),
+                      Column(
+                        mainAxisAlignment: MainAxisAlignment.center,
+                        children: const [
+                          Text(
+                            'Tổng đã chi',
+                            style: TextStyle(
+                                color: Colors.black,
+                                fontSize: 14,
+                                fontWeight: FontWeight.bold),
+                          ),
+                          SizedBox(height: 4.0),
+                          Text(
+                            '100.000đ',
+                            style: TextStyle(color: Colors.black),
+                          ),
+                        ],
+                      )
+                    ],
+                  ),
+                )
+              ],
+            ),
+          ),
+        ],
+      ),
+    );
+  }
+}
